+++ railties/test/application/middleware/remote_ip_test.rb
@@ -12,7 +12,9 @@ module ApplicationTests
       remote_ip = nil
       env = Rack::MockRequest.env_for("/").merge(env).merge!(
         "action_dispatch.show_exceptions" => false,
-        "action_dispatch.key_generator"   => ActiveSupport::LegacyKeyGenerator.new("b3c631c314c0bbca50c1b2843150fe33")
+        "action_dispatch.key_generator"   => ActiveSupport::CachingKeyGenerator.new(
+          ActiveSupport::KeyGenerator.new("b3c631c314c0bbca50c1b2843150fe33", iterations: 1000)
+        )
       )
 
       endpoint = Proc.new do |e|